Moved source files to 'src' folder
