--- Customer.java
@@ -1,19 +1,0 @@
-
-public class Customer extends Person {
-
-	private Book[] = new book[5]
-	private int id
-	public boolean rent(Book book) {
-		
-	}
-	public boolean returnBook(Book book) {
-		
-	}
-	private addBook(Book book) {
-		
-	}
-	private removeBook(Book book) {
-		
-	}
-	
-}
