+++ src/Library.java
@@ -1,0 +1,21 @@
+import java.util.ArrayList;
+
+public class Library  {
+    private StockKeeper sk;
+    private ArrayList<Customer> costumers = new ArrayList<Customer>();
+    private ArrayList<Librarian> librarians = new ArrayList<Librarian>();
+
+    Library(){}
+
+    public boolean rent(Book book){
+      //TODO Implement rent method
+    }
+
+    public boolean returnBook(Book book) {
+      //TODO Implement returnBook method
+    }
+
+    public void addBook(Book book) {}
+
+    public void removeBook(Book book) {}	
+}
